requirement.txt
amqp==2.5.2
billiard==3.6.2.0
boto3==1.12.6
botocore==1.15.6
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.10
django-appconf==1.0.3
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-filter==2.2.0
django-imagekit==4.0.2
django-redis==4.11.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
elasticsearch==7.5.1
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==0.9.5
kombu==4.6.7
MarkupSafe==1.1.1
openapi-codec==1.3.2
openpyxl==3.0.3
pilkit==2.0
Pillow==7.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
qiniu==7.2.6
qrcode==6.1
redis==3.4.1
requests==2.23.0
s3transfer==0.3.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.0
ujson==1.35
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0


